simple-markdown: ignore <br>s ending of list items
Summary:
This is almost purely for marked.js compatibility :(.
marked.js ignores double spaces (<br>s) at the end
of list items, so we are too for now. This is mostly
useful for testing that we don't differ from marked.js,
but there also seem to be a lot of items that have these
trailing spaces, and this is also a bit useful to not
break those when we launch this.

Test Plan: `make test`

Reviewers: emily

Reviewed By: emily

Differential Revision: http://phabricator.khanacademy.org/D13210