One weird trick you probably don't want me to do [md-parsing]
Summary:
Lists with items containing ' - ' were turning into lists with
sublists with items in them, even though ' - ' was mid-paragraph.

This commit adds hacks to get around the fact that we don't have
a nice block vs inline distinction yet. I intend to add that in
the future, so that lists will only work at a block level, and
then we can avoid all these hacks :).

But in the mean time I want to ship this so that I don't break
items on the site.

Test Plan:
`make test`

Auditors: emily