Two or more characters you should never type in a comment def
Summary:
Well, any two characters, to be specific. I forgot a `+` in a
regex and defs only ignored trailing defs if the trailing def
had exactly one characer in the [definition name].

I also changed a `+` to a `*` so that links to `<>` go to ""
instead of "<".

Test Plan:
`make test`

Auditors: emily