One string that is definitely not a tag! probably.
Summary:
Adds an ApiOption for a "group metadata editor" which will
allow us to embed a tagger widget to tag groups inside
perseus items.

Specifying a custom `apiOption.GroupMetadataEditor` will allow
clients to supply a full React class that can handle rendering
and editing the metadata for a group in whatever way it
would like. We'll build a GroupTagEditor in webapp and inject
that into our perseus editor there so that content creators
can edit tags.

To do this I had to add threading of the apiOptions prop through
the various editors we have on the editor side of the page in
addition to the renderer side of the page.

Test Plan:
 * typed "hi" in the metadata field in test.html
 * used developer json mode to see that this string was
   persisted as the metadata.

Auditors: alex