Fix orderer to work with React 0.12
props.ref disappeared, so my workaround for facebook/react#1392 broke. For now, just pass an extra prop that's equal to the ref so we can compare. This should go away after the root fix is merged in.

Test Plan: Dragged lots of cards to the beginning of an orderer, got no exceptions thrown.

Auditors: emily