SAT styling for radio buttons and checkboxes
Test Plan:
Load up an SAT quiz and check out questions with radio buttons or checkboxes.
Artificially change the font size of the katex and see that things still look ok.
Recognize that we need to flip a switch that enables radio buttons to be de-checked for SAT.
Dance party.

Reviewers: johnsullivan, aria

Subscribers: dylan

Differential Revision: http://phabricator.khanacademy.org/D14099