Don't send ALL THE JSON with interaction widget attempts
Summary: Oh, but maybe we should go back to sending something in other cases... oops...

Test Plan:
Locally, add an interaction widget and hit the "score" button.
See the guess for the interaction serialized as an empty object.
Add another widget, enter an answer, and hit "score".
Notice that the answer *is* actually serialized and not just undefined.

Reviewers: aria

Reviewed By: aria

Subscribers: joel, michelle, alex

Differential Revision: http://phabricator.khanacademy.org/D14128