Add background image support to grapher widget
Summary:
Added the ability to add background images to grapher widgets in
the same way that you can add them to interactive graphs. Also
fixed a bug in the grapher widget so it now gives an error when
you have too many ticks and abstracted the error-styling from the
number-line widget to be used in these error messages as well.

Test Plan:
- Add a background image to an item with a grapher widget
- Change the grid step of the grapher widget to 0.5 and note
  the red error message that appears in the editor.

Reviewers: aria

Reviewed By: aria

Maniphest Tasks: T3030

Differential Revision: http://phabricator.khanacademy.org/D14130