Adds callback to restore serialized state.
Summary:
restoreSerializedState changes the state of various react components
and so is not synchronous. Previously there was no way to tell when the
operation was complete. Now a callback is fired once the state has been
fully restored.

Test Plan:
Ran through various items with groups, passages, and simple radio
buttons and ensured that the state had been fully restored before the
callback was fired.

Reviewers: aria, alex

Reviewed By: alex

Subscribers: #sat, eater, dylan

Differential Revision: http://phabricator.khanacademy.org/D14604