Show fractional answers in review mode.
Summary:
This only affects the numeric input widget, and not the input number
widget because I couldn't find any misbehaving examples of the input
number widget to test with in any of the four SAT practice tests. Per
my meeting with Aria and Alex it also seems like we're not sure of the
direction we're going in with review mode so I don't want to implement
any more than necessary right now.

Before:
{F78782}

After:
{F78783}

Test Plan:
Found a widget that displayed the answer improperly in review mode,
then viewed it before and after this change.

Reviewers: alex, aria

Reviewed By: aria

Subscribers: eater, #sat

Differential Revision: http://phabricator.khanacademy.org/D14742