Adds ability to deselect radio buttons.
Summary:
Radio widgets can now be told to enable radio deselection which will
let users click the radio button again to deselect them.

Aria and Alex mentioned that they were worried about this behavior not
being discoverable. Kitt has argued that he doesn't actually see this
being an issue as long as there are contextual clues (ie: you're taking
a test where you might want to deselect an answer). Since this will
supposedly only be enabled in a test taking environment, this seems to
make sense.

To address discoverability in unforeseen cases where this sort of
context isn't really available, we could make some additive changes to
this that would annotate a selected answer with a clue like "click
again to deselect". This could be done as simply as adding a CSS class
to affected radio buttons.

That being said, I'm interested in hearing anyone's thoughts on this.
We really want this functionality to be present in some form by Dec 1
though, even if just as a temporary thing that we pull out immediately
after. This is because of the large number of students who will be
using the SAT product, and the fact that a wrong answer is penalized
more than just not selecting an answer.

{F79370}

Test Plan:
* Tested in Perseus editor (localhost:9000/test.html) to verify that
  things look sane from a content creators viewpoint.
* Tested in the SAT section viewer to make sure the experience is
  reasonable for the test taker (forced the option to be enabled with
  hacks until we can run scripts to enable this option on all the past
  questions).

Reviewers: alex, aria

Reviewed By: aria

Subscribers: kitt, #sat, eater

Differential Revision: https://phabricator.khanacademy.org/D14815