[Versioning] Don't traverse widget props you don't understand
Summary:
We can't do reasonable things with widget editor props during a
widget traversal (here for the context of checking isRenderable
on an item) if the widget is at a version later than this
of perseus supports. In this case, we must not recurse on those
widget's props.

This sounds scary, but is safe in the context of versioning because
it just means that if we find a widget we don't understand,
we will mark the whole item as incompatible, and it doesn't
matter whether we traverse that widget's children or not.

If future use cases need to be able to traverse the children of
widgets they don't understand, they should stop asking to do the
impossible.

Depends on D16248

Test Plan: Unit tests added next commit

Reviewers: alex, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D16249