Adds a "group aware" get widget IDs function.
Summary:
This is meant as a temporary (or at least experimental) hack to allow
the SAT section viewer to determine if questions are multiple choice or
not. This is necessary because for the SAT, incorrect multiple choice
questions are penalized with -0.25 points, whereas incorrect numeric
input questions are not penalized.

Aria mentioned that she's working on something that will hopefully open
a more permanent way to do this. Thus why this is intended to be
temporary.

Test Plan:
* Created an item with three groups, one of them nested in another, then
  executed `getAllWidgetIds()` and checked the result.
{F79421}
* Tried it on an item with no widgets.

Reviewers: alex, aria

Reviewed By: aria

Subscribers: #sat, eater

Differential Revision: https://phabricator.khanacademy.org/D14825