Make SvgImage respect scale
Summary:
Make SvgImage render SVGs and the labels on top of them with
the specified scales. In particular, the labels are rendered on the
appropriate scale but their font size is increased to match the scale.

This is used in the editor for interactive-graph, grapher, etc.
Previously, if a graphie-to-svg background was used, the image in the
editor would be too big. Now it should be the correct size.

Test Plan:
 - Create interactive-graph and grapher widgets.
 - Add a graphie-to-svg image as the background image for both. (For
   example:
   `web+graphie://ka-perseus-graphie.s3.amazonaws.com/991e6d9b763e133c59a94010e8c308ef2ff78757`)
   - Ensure that the graph in the body of the question is the normal
     size.
   - Ensure that the graph in the editor is smaller, and the same size
     as the graph that it replaced.
 - Add a normal image as the background image for both. (For example:
   `http://ka-perseus-graphie.s3.amazonaws.com/009bde8c529cd0f9c8fea990aac9937603b26526.png`)
   - Do the same two things as above.

Reviewers: alex, aria

Reviewed By: aria

Differential Revision: https://phabricator.khanacademy.org/D14974