Don't calculate graphie attributes before data is loaded
Summary:
When the label/other data hadn't been loaded, and a size for
the image wasn't provided (like what happens the first time an image is
placed into a markdown area), we still tried to calculate the graphie
`box` attribute from the (unloaded) data, which caused an error. Now, we
only calculate graphie attributes when the data is loaded.

Test Plan:
 - Add a web+graphie markdown link into the body of the question.
   - See that it renders correctly.
   - See that there are no JS errors.
 - Add a (different) web+graphie markdown link into a choice in a
   multiple-choice widget.
   - See that it renders correctly.
   - See that there are no JS errors.

Reviewers: alex, aria

Reviewed By: alex, aria

Differential Revision: https://phabricator.khanacademy.org/D15044