Change the way widget shortcuts work
Summary:
Previously, widget shortcuts had to be an exact shortcut sequence and
could be completed with space, tab, or enter. This diff makes it so that
shortcuts are like command-line tab completion (press TAB to insert the
widget if you have typed enough characters to uniquely specify a
widget).

Depends on D15230.

Test Plan:
Try typing `[[categ`-TAB and see that inserts a categorizer widget
Check that you can still add them from the dropdown

Reviewers: michelle

Reviewed By: michelle

Differential Revision: https://phabricator.khanacademy.org/D15231