[Radio] Allow contributors to mark correct choice as 'none of the above'
Currently in order to have the correct choice in the Radio widget to be hidden as 'None of the above', it must happen to be chosen randomly via the 'Auto-none of the above' feature. With this commit the contributor has now the ability to mark a correct choice as 'none of the above' itself if he/she so desire.