Expose HintsRenderer to the outside world.
Summary:
This change pulls the HintsRenderer component out into its own file,
and exposes it so that you can use it outside of Perseus. This is
the same way the Renderer component is exposed.

This is the result of some thought over whether we should refactor the
SAT Mission to use the ItemRenderer, or just leave it in its current
state of using the Renderer directly, and then doing this.

The ItemRenderer doesn't conform with the ways that the SAT Mission
has been structured so far. Namely we have no answer area at all, and
we use Perseus-style scores throughout. These can be overcome, but it
is not clear us using the ItemRenderer directly is better for Perseus
in the longterm. It may make sense to instead make a different kind of
ItemRenderer inside or outside of Perseus. This has basically been done
already in the form of the SatPerseusRenderer in
`sat-perseus-renderer.jsx` in the webapp repo.

I am very open to discuss this and go a different route, and do not
feel super opinionated that using the HintsRenderer is the way to go.
Interested in other opinions here.

Test Plan:
* Created a HintsRenderer component outside of Perseus. To do this, you
  can go to `sat-perseus-renderer.jsx`, add the line
  `<Perseus.HintsRenderer hints={[]} />` in the `render()` function's
  multi-line return statement (try between the span and
  Perseus.Renderer components). Then visit /mission/sat/tasks/123. If
  you see no javascript errors, things are good.
* Went to /math/trigonometry/less-basic-trigonometry/angle-addition-formulas-trig/e/trig_addition_identities
  and clicked "I'd like a hint", and saw a hint appear.
* make test
* Ran `make server` (in the Perseus repo). Went to
  localhost:9000/test.html and created a hint.

Reviewers: alex, aria

Reviewed By: aria

Subscribers: #sat

Differential Revision: https://phabricator.khanacademy.org/D15877