[Makefile] delete react-components node_modules in `make install`
Summary:
Having node_modules lower in the tree is scary for both
compatibility and the size of the output perseus.js file.

In this case, react-components' React was incompatible with
our test cases. Next time, it could have been a React that
output different types of elements, or a duplicate
underscore. This is scary. Let's not have this

Test Plan:
tscrossedfingers

Auditors: emily