Prevent orderer from getting stuck on touch devices
Summary:
A touch event, it seems, includes a `changedTouches` property and a `touches`
property. It seems that the `changedTouches` includes touches that started on
the bound-to element; `touches` includes all touches currently in play.

What we seem to be doing here is tracking one touch at a time via `pointerDown`
and `currentTouchIdentifier`. I.e., the pointer is either down or not. This
isn't a terrible idea: We'll look for any touchdown, but then are careful to
only respond to its corresponding touchup.

The problem is that we were grabbing the touchdown from `event.touches` and
looking for the touchup on `event.changedTouches`. The former might include
touches that are outside the bound-to element (e.g., if there's a two-finger
touch where one finger is in the bound-to element and one finger is outside,
`event.touches[0]` might be the finger that's outside). If that happens, we'll
never see the corresponding touchup on `changedTouches`, thus ignoring any
subsequent touches. :(

This change looks only at `changedTouches` to ensure that all touches we
consider definitely start within the bound-to element.

This diff also includes a check that `this.refs.placeholder` is defined.
I ran into an unrelated problem where sometimes it wasn't and the JS error
would result in the card getting stuck on mouseup halfway between the stack and
the sorter box (though it could be unstuck by trying to drag it again, so not
catastrophic)

Test Plan:
1. Point the iOS app at local webapp instance with this perseus change.
2. From iOS app, load up "Counting with small numbers" exercise
3. Two-finger tap on the stack of cards with one finger on the stack and the other finger outside the stack. Repeat this a few times since it seems like there's a 50/50 chance of either finger being listed first in the touches property of the event.
4. Drag things back and forth, answer questions normally, etc.
5. Do lots of other crazy dragging motions to verify cards don't get stuck anywhere.
6. Load the "Counting with small numbers" in webapp and verify it still works normally on the web.

Reviewers: emily

Reviewed By: emily

Subscribers: aria, alan

Differential Revision: https://phabricator.khanacademy.org/D16431