[radio] i18n "None of the above"
Summary:
In the past, our "None of the above" user-facing strings have
not been translated. This isn't as bad as it sounds because
only a couple exercises in CS used them, but it is still bad.

This commit fixes that!

Test Plan:
No real test plan for the translation, but I ran `make build`
and looked through `build/perseus-3.js` to see whether these
instances are wrapped in `$_` functions correctly, and they
were.

I also made a radio in an item, added a correct "none of the above"
choice, and left it's 'reveal' answer empty, and answered
it, so that I could see both "None of the above" instances
displayed on screen.

Auditors: eater