[radio] Always display "None of the above" with a Renderer
Summary:
This makes the output of the two code paths that can display
"None of the above" to the user hopefully the same. There isn't
a visual change right now, but for consistency this makes
more sense to me.

Test Plan:
Made a radio in an item, added a correct "none of the above"
choice, and left it's 'reveal' answer empty, and answered
it, so that I could see both "None of the above" instances
displayed on screen.

Auditors: eater